BrainCoEEG/requirements.txt
pandas
numpy==1.26.4
matplotlib
scipy
seaborn
pyyaml
scikit-learn
joblib
bc_device_sdk==0.3.5
asyncio>=3.4.3
